+++ src/Atributos/DecoradorDano.java
@@ -1,6 +1,6 @@
 package Atributos;
 
 public abstract class DecoradorDano {
-    public abstract int getValue(int i);
-    public abstract String toString(int modificador);
+    public abstract float getValue(float i);
+    public abstract String toString(float modificador);
 }
